requirements.txt
paho-mqtt==1.6.1
pymorphy2==0.9.1